+++ libraries/exoplayer_rtsp/src/main/java/androidx/media3/exoplayer/rtsp/RtspClient.java
@@ -337,19 +337,22 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
   }
 
   /**
-   * Gets the included {@link RtspMediaTrack RtspMediaTracks} from a {@link SessionDescription}.
+   * Returns the included {@link RtspMediaTrack RtspMediaTracks} from parsing the {@link
+   * SessionDescription} within the {@link RtspDescribeResponse}.
    *
-   * @param sessionDescription The {@link SessionDescription}.
+   * @param rtspDescribeResponse The {@link RtspDescribeResponse} from which to retrieve the tracks.
    * @param uri The RTSP playback URI.
    */
   private static ImmutableList<RtspMediaTrack> buildTrackList(
-      SessionDescription sessionDescription, Uri uri) {
+      RtspDescribeResponse rtspDescribeResponse, Uri uri) {
     ImmutableList.Builder<RtspMediaTrack> trackListBuilder = new ImmutableList.Builder<>();
-    for (int i = 0; i < sessionDescription.mediaDescriptionList.size(); i++) {
-      MediaDescription mediaDescription = sessionDescription.mediaDescriptionList.get(i);
+    for (int i = 0; i < rtspDescribeResponse.sessionDescription.mediaDescriptionList.size(); i++) {
+      MediaDescription mediaDescription =
+          rtspDescribeResponse.sessionDescription.mediaDescriptionList.get(i);
       // Includes tracks with supported formats only.
       if (RtpPayloadFormat.isFormatSupported(mediaDescription)) {
-        trackListBuilder.add(new RtspMediaTrack(mediaDescription, uri));
+        trackListBuilder.add(
+            new RtspMediaTrack(rtspDescribeResponse.headers, mediaDescription, uri));
       }
     }
     return trackListBuilder.build();
@@ -618,7 +621,9 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
           case METHOD_DESCRIBE:
             onDescribeResponseReceived(
                 new RtspDescribeResponse(
-                    response.status, SessionDescriptionParser.parse(response.messageBody)));
+                    response.headers,
+                    response.status,
+                    SessionDescriptionParser.parse(response.messageBody)));
             break;
 
           case METHOD_SETUP:
@@ -708,7 +713,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
         }
       }
 
-      ImmutableList<RtspMediaTrack> tracks = buildTrackList(response.sessionDescription, uri);
+      ImmutableList<RtspMediaTrack> tracks = buildTrackList(response, uri);
       if (tracks.isEmpty()) {
         sessionInfoListener.onSessionTimelineRequestFailed("No playable track.", /* cause= */ null);
         return;