+++ libraries/exoplayer_rtsp/src/main/java/androidx/media3/exoplayer/rtsp/RtspDescribeResponse.java
@@ -20,6 +20,8 @@ import androidx.media3.common.util.UnstableApi;
 /** Represents an RTSP DESCRIBE response. */
 @UnstableApi
 /* package */ final class RtspDescribeResponse {
+  /** The response's headers. */
+  public final RtspHeaders headers;
   /** The response's status code. */
   public final int status;
   /** The {@link SessionDescription} (see RFC2327) in the DESCRIBE response. */
@@ -28,10 +30,13 @@ import androidx.media3.common.util.UnstableApi;
   /**
    * Creates a new instance.
    *
+   * @param headers The response's headers.
    * @param status The response's status code.
    * @param sessionDescription The {@link SessionDescription} in the DESCRIBE response.
    */
-  public RtspDescribeResponse(int status, SessionDescription sessionDescription) {
+  public RtspDescribeResponse(
+      RtspHeaders headers, int status, SessionDescription sessionDescription) {
+    this.headers = headers;
     this.status = status;
     this.sessionDescription = sessionDescription;
   }